refactor(listener): extract connection lifecycle
